fix: create tmp dir if it does not exist

+++ demos-core/KendoCRUDService/KendoCRUDService/Program.cs
@@ -40,12 +40,24 @@ if (builder.Environment.IsProduction())
 
 
 var dbPath = Path.Combine(builder.Environment.ContentRootPath, "wwwroot", "data", "sample.db");
-var tmpDbPath = Path.Combine("tmp", "demos-local.db");
+var tmpDir = Path.Combine(builder.Environment.ContentRootPath, "tmp");
+var tmpDbPath = Path.Combine(tmpDir, "demos-local.db");
 
 if (builder.Environment.IsProduction())
 {
     // In production, the SQLite DB is copied to a temp location to avoid locking issues.
-    File.Delete(tmpDbPath);
+    if (!Directory.Exists(tmpDir))
+    {
+        Console.WriteLine($"Creating directory: {tmpDir}");
+        Directory.CreateDirectory(tmpDir);
+    }
+
+    if (File.Exists(tmpDbPath))
+    {
+        Console.WriteLine($"Deleting existing file: {tmpDbPath}");
+        File.Delete(tmpDbPath);
+    }
+
     File.Copy(dbPath, tmpDbPath);
     dbPath = tmpDbPath;
 }
